requirements.txt
scipy
scikit-learn
gym
mujoco-py
matplotlib
tensorflow-gpu 
tensorflow-probability
tqdm
